Port Array `with` from FM-Core to FM-Lang
`with` is the most important Array function, all the others are easily
derived from it. It updates an element in an index and returns the old
element (optionally changed). Writing this was difficult because, in
FMC, it was implemented in a tail-recursive, bottom-up manner. If I
implemented it normally, its type would be, simply:

    with
    : {D : Rec} ->
    ! { path : U32
      , fun  : {x : U32} -> U32
      , arr  : [el : U32, Array(to_nat(D))]
      }     -> [el : U32, Array(to_nat(D))]

That is, "with" receives a path, an array, updates an element and
returns the old. This is prety easy to implement, but will cause indices
to be scrambled. To implement it bottom-up, instead of passing the
*array* down through the recurson, we must instead pass a *function*
that receives the array and performs the desirable update, and build it
recursivelly.

    {D : Rec} -> Type
    { ~n   : Nat
    , path : U32
    , fun  : {x : Array(n)} -> [el : U32, Array(n)]
    , arr  : Array(open<Nat>(add_from_rec(D, #n)))
    }     -> [el : U32, Array(open<Nat>(add_from_rec(D, #n)))]

This requires keeping track of the size of the array the function deals
with. On the topmost part of the recursive tree, we will have a function
`fn` that works on `Array(0)`. After the first recursive call, `fn`
works on `Array(1)`. And so on, until we've got a function that works in
our input array, `Array(to_nat(D))`. Problem is, we need to convince the
type-checker that, after N calls, the `N` we get is equivalent to
`to_nat(D)`, which requires a bunch of rewrites, and a proof that:

    ∀ (r : Rec) (n : Nat) -> r + succ(n) == call(r) + n

This commit correctly annotates a `with` function that erases to the old
FMC one. This implementation is non-polymorphic (only U32 elements), but
should be easy to change now.