Emulate structural recursion with Bound type
Consider the following proof:

    // ∀ n . n < n+1
    !less_than_succ*N : !{case n : Nat} -> Less(n, succ(n))
    | succ => less_succ(~n.pred, ~succ(n.pred), less_than_succ(n.pred))
    | zero => less_zero(~zero)
    * ?

As is, it can't be completed since, on the halt-case, we'll need to
prove `Less(n, succ(n))`, which we can't, since that's what we're trying
to prove. In other words, the fact Formality recursive functions need a
halt case makes that proof impossible.

We could try, for example, returning `Maybe(Less(n, succ(n)))`. That'd
be useful in the sense that "if we call the function and it doesn't
return nothing, then our theorem holds". But the fact the function can
always return `nothing` anytime invalidates it as a proof in the
mathematical sense:

    // ∀ n . n < n+1
    !less_than_succ*N : !{case n : Nat} -> Maybe(Less(n, succ(n)))
    | succ => nothing
    | zero => nothing
    * nothing

Similarly, we could, in some clever way, return a proof that either
`Less(n, succ(n))`, or that the function ran out of gas". But that still
accomplishes nothing, since we could just force the function to run out
of gas:

    // ∀ n . n < n+1
    !less_than_succ*N : !{case n : Nat} -> TrueIfHalted(N, Less(n, succ(n)))
    | succ => less_than_succ(zero)
    | zero => less_than_succ(zero)
    * did_not_halt

In order to solve it in a way that preserves the mathematical meaning of
the proof, my idea is to emulate Agda/Coq's structural recursion with a
`Bound` type. That type compares the size of the input (here, a `Nat`)
with the amount of recursion calls we have left; i.e., our "fuel";,
asserting that we always "have enough fuel" to consume the whole input.
The proof then goes like ths:

    T Bound (n : Nat, i : Ind)
    | bound_succ {~n : Nat, ~i : Ind, k : Bound(n, i)} (succ(n), step(i))
    | bound_zero {~i : Ind}                            (zero   , step(i))

    // ∀ n . n < n+1
    !less_than_succ*N : !{n : Nat, bound : Bound(n, N)} -> Less(n, succ(n))
      (case/Bound bound
      | bound_succ => {e}
        let r0 = cong_unstep(~i, ~N, ~e)
        let r1 = less_than_succ(n, k :: rewrite x in Bound(n, x) with r0)
        let r2 = less_succ(~n, ~succ(n), r1)
        r2
      | bound_zero => {e}
        less_zero(~zero)
      : {e : i == step(N)} -> Less(n, succ(n)))(refl(~step(N)))
      * absurd(absurd_bound(n, bound), ~Less(n, succ(n)))

The insight is that, instead of matching on `n` directly, we match on
`bound`, which will give us not only the predecessor of `n`, `n - 1`,
but also the predecessor of the call count, `N - 1`, together with a
proof that it is still larger the number.  We can then use this new
proof to recurse on `n`'s predecessor. That guarantees that `n` not only
always gets smaller, but that it will reach `zero` before our "call
count" reaches its limit, essentially emulating Agda's bounded
recursion. This Bound only works for Nats, but should generalize.

Note: this proof is noisy since it requires us to convince the
type-checker that the `N` inside the bound field is the same as the `N`
from the function input, which, in turn, required me to prove a bunch of
`Ind` theorems. Agda has a built-in syntax that gives you this equality
for free, I believe. Should we too?