Began refactor of Algebra hierarchy based on Setoids
Previously algebraic structures looked like:

```
Associative : {A : Type, f : A -> A -> A} -> Type
   {x : A, y : A, z : A} -> f(f(x,y),z) == f(x,f(y,z))

Semigroup {A : Type}
| semigroup
  { f           : A -> A -> A
  , associative : Associative(A,f)
  }
```

This was fine, but made it impossible to define, e.g. the Semigroup of
endomorphisms `A -> A`, since we have no way of constructing `==` for
functions.

Now the algebraic structures look like

```
Associative : {A : Type, s : Setoid(A), f : Op2(A)} -> Type
  case/Setoid s | setoid =>
    {x : A, y : A, z : A} -> s.R(f(f(x,y),z), f(x,f(y,z)))
  : Type

T Semigroup {A : Type, s : Setoid(A)}
| semigroup
  { f           : Op2(A)
  , associative : Associative(A,s,f)
  }
```

This is much more verbose, but does enable us to define extentional
structures:

```
semigroup.endo : {~A : Type} -> Semigroup(A -> A)
  let f = compose(~A,~A,~A)
  let s = setoid(~(A -> A), Feq(A,A), feq.equivalence(~A,~A))
  let associative = {x,y,z,a} refl(~x(y(z(a))))
  semigroup3(~(A -> A),f,~s,associative)
```

Also in this commit are refactors to `Relation.Binary` and
`Relation.Equality` to support the Setoid refactor